+++ src/main/java/subway/line/domain/Section.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.domain;
 
 public class Section {
     private Long id;
