+++ src/main/java/subway/line/domain/Sections.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.domain;
 
 import java.util.List;
 import java.util.Map;
