+++ src/main/java/subway/line/service/SectionService.java
@@ -1,10 +1,13 @@
-package subway.line;
+package subway.line.service;
 
 import org.springframework.stereotype.Service;
 import subway.exception.DeleteSectionException;
 import subway.exception.SectionDistanceExceedException;
-import subway.station.Station;
-import subway.station.StationDao;
+import subway.line.domain.Section;
+import subway.line.dao.SectionDao;
+import subway.line.domain.Sections;
+import subway.station.domain.Station;
+import subway.station.dao.StationDao;
 
 import javax.annotation.Resource;
 import java.util.List;
