+++ src/main/java/subway/station/controller/StationController.java
@@ -1,11 +1,15 @@
-package subway.station;
+package subway.station.controller;
 
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.line.LineService;
-import subway.line.SectionService;
+import subway.line.service.LineService;
+import subway.line.service.SectionService;
+import subway.station.domain.Station;
+import subway.station.domain.StationRequest;
+import subway.station.domain.StationResponse;
+import subway.station.service.StationService;
 
 import javax.annotation.Resource;
 import java.net.URI;
