+++ src/main/java/subway/station/dao/StationDao.java
@@ -1,9 +1,11 @@
-package subway.station;
+package subway.station.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import subway.station.domain.Station;
+import subway.station.util.StationMapper;
 
 import javax.annotation.Resource;
 import java.sql.PreparedStatement;
