+++ src/main/java/subway/station/domain/StationRequest.java
@@ -1,4 +1,4 @@
-package subway.station;
+package subway.station.domain;
 
 public class StationRequest {
     private String name;
