+++ src/test/java/subway/line/SectionAcceptanceTest.java
@@ -9,7 +9,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import subway.AcceptanceTest;
-import subway.station.StationResponse;
+import subway.line.domain.LineResponse;
+import subway.line.domain.SectionRequest;
+import subway.station.domain.StationResponse;
 
 import java.util.Arrays;
 import java.util.List;
